refactor: separate lookup classes into files
