+++ twaddle/interpreter/interpreter.py
@@ -4,7 +4,8 @@ from re import Match, sub
 
 from twaddle.exceptions import TwaddleInterpreterException
 from twaddle.interpreter.formatter import Formatter
-from twaddle.lookup.lookup import LookupDictionary, LookupManager
+from twaddle.lookup.lookup_manager import LookupManager
+from twaddle.lookup.lookup_dictionary import LookupDictionary
 from twaddle.parser.compiler import Compiler
 from twaddle.parser.compiler_objects import (
     BlockObject,
